requirements.txt
flask
flask-socketio
cryptography
pycryptodome
websockets
requests
jsonpickle
